+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### 2.3.0 TBD
+- Change: A TUN-device is used instead of firewall rules for routing outbound connections.
+- Change: Outbound connections now use gRPC instead of ssh, and the traffic-manager no longer has a sshd running.
+
 ### 2.2.2 (May 17, 2021)
 - Feature: Telepresence translates legacy Telepresence commands into viable Telepresence commands.
 - Bugfix: Intercepts will only look for agents that are in the same namespace as the intercept.